+++ docs/Feedback.md
@@ -2,9 +2,9 @@
 
 ## 1단계 - HTTP 서버 구현하기 1차 피드백
 
-- [ ] queryParam 일급 컬렉션 만들고 객체 생성 책임 분리
+- [x] queryParam 일급 컬렉션 만들고 객체 생성 책임 분리
     - [ ] QueryStringHandler 사용 목적
-    - [ ] QueryStringHandler 매직넘버
+    - [x] QueryStringHandler 매직넘버
 - [ ] MyHttpResponse 메서드 수정
     - [ ] String으로 변경해주는 메서드 생성
-- [ ] Http11Processor - 불필요한 `;` 제거
+- [x] Http11Processor - 불필요한 `;` 제거
