+++ tomcat/src/main/java/org/apache/coyote/domain/HttpRequest.java
@@ -1,32 +1,27 @@
 package org.apache.coyote.domain;
 
-import java.util.HashMap;
-import java.util.Map;
-import org.apache.coyote.handler.QueryStringHandler;
-
 public class HttpRequest {
 
     private static final String HEADER_DELIMITER = " ";
 
     private final String uri;
-    private final Map<String, String> queryParam;
+    private final QueryParam queryParam;
 
-    private HttpRequest(String uri, Map<String, String> queryParam) {
+    private HttpRequest(String uri, QueryParam queryParam) {
         this.uri = uri;
         this.queryParam = queryParam;
     }
 
     public static HttpRequest from(String headerFirstLine) {
         String[] headers = headerFirstLine.split(HEADER_DELIMITER);
-        Map<String, String> queryParam = QueryStringHandler.queryParams(headers[1]);
-        return new HttpRequest(headers[1], queryParam);
+        return new HttpRequest(headers[1], QueryParam.from(headers[1]));
     }
 
     public String getUri() {
         return uri;
     }
 
-    public Map<String, String> getQueryParam() {
+    public QueryParam getQueryParam() {
         return queryParam;
     }
 }
